fix(tests): #4503 Refactorred filtering tests.
